artist/all working - about to be added, artist/[artist_id] page done
